+++ test/test_vendor_hiveeyes.py
@@ -8,7 +8,7 @@ from bunch import Bunch
 
 from test.conftest import create_machinery
 from test.resources import PROCESS_DELAY
-from test.util import mqtt_sensor, sleep, InfluxWrapper, GrafanaWrapper
+from test.util import mqtt_json_sensor, sleep, InfluxWrapper, GrafanaWrapper
 
 logger = logging.getLogger(__name__)
 
@@ -45,7 +45,7 @@ def test_mqtt_to_grafana(machinery_hiveeyes, create_influxdb_hiveeyes, reset_inf
         'temperature': 42.84,
         'weight': 33.33,
     }
-    yield mqtt_sensor(settings.mqtt_topic, data)
+    yield mqtt_json_sensor(settings.mqtt_topic, data)
 
     # Wait for some time to process the message.
     yield sleep(PROCESS_DELAY)